bombのディレクトリを切った
