+++ src/bomb/component.rs
@@ -1,0 +1,5 @@
+use bevy::prelude::*;
+
+// BombのComponent
+#[derive(Component)]
+pub struct Bomb;
